Add files via upload
This helps maintaining the patients history and their past medical records. Allows them to have history uploaded automtically just by reading the excel and maintain the amount and history of  a patient.

Hospita agents can send the attachment by submitting a form or by sending an email. It reads the excel with the patient name and Patient Unique medical number and updates the information and maintain the history