Add files via upload
This helps maintaining the patients history and their past medical records. Allows them to have history uploaded automtically just by reading the excel and maintain the amount and hisory of  a patient

This maintain the history and calcualtes the patient total bill from admitted date till discharged date and maintain the data with patients history
Hospital's agent can either send an email or submits the form from the portal. It read the whole excel and populates the information in Patient Table records and maintain the data based on Patient's name and Patient's unqieu medical number.

It helps each origanization to maintain the data and see the patient's history as well. Also, helps in pulling the report on monthly basis